+++ module/src/main/java/com/iluwatar/module/App.java
@@ -21,64 +21,75 @@ package com.iluwatar.module;
 import java.io.FileNotFoundException;
 
 /**
- * The Module pattern can be considered a Creational pattern and a Structural
- * pattern. It manages the creation and organization of other elements, and
- * groups them as the structural pattern does. An object that applies this
- * pattern can provide the equivalent of a namespace, providing the
- * initialization and finalization process of a static class or a class with
- * static members with cleaner, more concise syntax and semantics.
+ * The Module pattern can be considered a Creational pattern and a Structural pattern. It manages
+ * the creation and organization of other elements, and groups them as the structural pattern does.
+ * An object that applies this pattern can provide the equivalent of a namespace, providing the
+ * initialization and finalization process of a static class or a class with static members with
+ * cleaner, more concise syntax and semantics.
  * <p>
- * The below example demonstrates a use case for testing two different modules:
- * File Logger and Console Logger
+ * The below example demonstrates a use case for testing two different modules: File Logger and
+ * Console Logger
  * 
  */
 public final class App {
 
-	public static FileLoggerModule fileLoggerModule = null;
-	public static ConsoleLoggerModule consoleLoggerModule = null;
+  public static FileLoggerModule fileLoggerModule = null;
+  public static ConsoleLoggerModule consoleLoggerModule = null;
 
-	public static void prepare() throws FileNotFoundException {
+  /**
+   * Following method performs the initialization
+   * 
+   * @throws FileNotFoundException if program is not able to find log files (output.txt and
+   *         error.txt)
+   */
+  public static void prepare() throws FileNotFoundException {
 
-		fileLoggerModule = FileLoggerModule.getSingleton();
-		consoleLoggerModule = ConsoleLoggerModule.getSingleton();
+    fileLoggerModule = FileLoggerModule.getSingleton();
+    consoleLoggerModule = ConsoleLoggerModule.getSingleton();
 
-		/* Prepare modules */
-		fileLoggerModule.prepare();
-		consoleLoggerModule.prepare();
-	}
+    /* Prepare modules */
+    fileLoggerModule.prepare();
+    consoleLoggerModule.prepare();
+  }
 
-	public static void unprepare() {
+  /**
+   * Following method performs the finalization
+   */
+  public static void unprepare() {
 
-		/* Close all resources */
-		fileLoggerModule.unprepare();
-		consoleLoggerModule.unprepare();
-	}
+    /* Close all resources */
+    fileLoggerModule.unprepare();
+    consoleLoggerModule.unprepare();
+  }
 
-	public static final void execute(final String... args) {
+  /**
+   * Following method is main executor
+   * 
+   * @param args for providing default program arguments
+   */
+  public static void execute(final String... args) {
 
-		/* Send logs on file system */
-		fileLoggerModule.printString("Message");
-		fileLoggerModule.printErrorString("Error");
+    /* Send logs on file system */
+    fileLoggerModule.printString("Message");
+    fileLoggerModule.printErrorString("Error");
 
-		/* Send logs on console */
-		consoleLoggerModule.printString("Message");
-		consoleLoggerModule.printErrorString("Error");
-	}
+    /* Send logs on console */
+    consoleLoggerModule.printString("Message");
+    consoleLoggerModule.printErrorString("Error");
+  }
 
-	/**
-	 * Program entry point.
-	 * 
-	 * @param args
-	 *            command line args.
-	 * @throws FileNotFoundException
-	 */
-	public static final void main(final String... args)
-			throws FileNotFoundException {
-		prepare();
-		execute(args);
-		unprepare();
-	}
+  /**
+   * Program entry point.
+   * 
+   * @param args command line args.
+   * @throws FileNotFoundException if program is not able to find log files (output.txt and
+   *         error.txt)
+   */
+  public static void main(final String... args) throws FileNotFoundException {
+    prepare();
+    execute(args);
+    unprepare();
+  }
 
-	private App() {
-	}
+  private App() {}
 }
